+++ apps/documents/models.py
@@ -1,7 +1,6 @@
 import logging
 
 from django.conf import settings
-from django.core.validators import MinValueValidator, MaxValueValidator
 from django.db import models
 from enumfields import Enum
 from enumfields import EnumIntegerField
@@ -41,14 +40,14 @@ class Document(models.Model):
     id = models.CharField(primary_key=True, default=random_id, max_length=36)
     name = models.CharField(max_length=36, null=False, blank=False)
     description = models.CharField(max_length=250, null=True, blank=True)
-    uploaded_at = models.DateTimeField(auto_now_add=True)
     owner_id = models.CharField(null=False, max_length=36)
     document_type = EnumIntegerField(enum=DocumentType, default=DocumentType.BOL)
     file_type = EnumIntegerField(enum=FileType, default=FileType.PDF)
     upload_status = EnumIntegerField(enum=UploadStatus, default=UploadStatus.PENDING)
-    size = models.IntegerField(validators=[MinValueValidator(1), MaxValueValidator(12500000)])
     s3_path = models.CharField(max_length=252, blank=False, null=False)
     shipment = models.ForeignKey(Shipment, on_delete=models.CASCADE, null=False)
+    updated_at = models.DateTimeField(auto_now=True)
+    created_at = models.DateTimeField(auto_now_add=True)
 
     @property
     def pre_signed_url(self):
